pyproject.toml
[tool.poetry]
name = "bs-monitor"
version = "0.1.0"
description = "Sneaker monitor"
authors = ["Ryan Bradshaw <ryan@rbradshaw.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
httpx = "^0.16.1"
beautifulsoup4 = "^4.9.3"
PyYAML = "^5.3.1"
aiofiles = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
pytest-asyncio = "^0.14.0"
bpython = "^0.20"
black = "^20.8b1"
isort = "^5.6.4"
flake8 = "^3.8.4"

[tool.black]
line-length = 120
target-version = ['py38']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"